% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_exemplar.R
\name{cluster_exemplar}
\alias{cluster_exemplar}
\title{Cluster Exemplar}
\usage{
cluster_exemplar(cluster_no)
}
\arguments{
\item{cluster_no}{This is the cluster number, as an int.}
}
\description{
This function returns the cluster examplar as the gene name.
}
\examples{
cluster_exemplar(150)
}

