% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_exemplar_array.R
\name{cluster_exemplar_array}
\alias{cluster_exemplar_array}
\title{Cluster Exemplar Array}
\usage{
cluster_exemplar_array(cluster_list)
}
\arguments{
\item{cluster_list}{This is the list of cluster numbers.}
}
\description{
This function returns the an array of cluster examplars given the cluster numbers.
}
\examples{
cluster_exemplar_array([1:10])
}

