% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_number.R
\name{cluster_number}
\alias{cluster_number}
\title{Cluster Number}
\usage{
cluster_number(neuron_name)
}
\arguments{
\item{neuron_name}{This is the neuron name (i.e. "fru-M-200145")}
}
\description{
This function returns the cluster number that a particular neuron belongs to.
}
\examples{
cluster_number("fru-M-200145")
}

