% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_scan.R
\name{cluster_scan}
\alias{cluster_scan}
\title{Cluster Scan}
\usage{
cluster_scan(clusterlist = 1:809)
}
\arguments{
\item{clusterlist}{List of cluster numbers to scan through. Default is 1:809.}
}
\description{
This function scans through a list of clusters (integers), allowing the user to select clusters that will be returned as an array of integers.
}
\examples{
cluster_scan(1:10)
}

