% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmirrorneuron.R
\name{fcmirrorneuron}
\alias{fcmirrorneuron}
\title{Mirror neurons}
\usage{
fcmirrorneuron(fcneuron)
}
\arguments{
\item{fcneuron}{This is the name of the original neuron.}
}
\description{
This function returns the mirrored neuron. The bounding box is the FCWB.surf.
}
\examples{
fcmirrorneuron("fru-M-200145")
}

