% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcneuroncluster.R
\name{fcneuroncluster}
\alias{fcneuroncluster}
\title{Return neurons in a given cluster}
\usage{
fcneuroncluster(cluster_no)
}
\arguments{
\item{cluster_no}{This is the cluster number}
}
\description{
This function returns the neurons belonging to a given cluster as a neuronlist.
}
\examples{
fcneuroncluster(5)
}

