% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcneuronlist.R
\name{fcneuronlist}
\alias{fcneuronlist}
\title{Return neurons}
\usage{
fcneuronlist(fcneuronnames)
}
\arguments{
\item{fcneuronnames}{This is the name of the neuron (i.e. "fru-M-200145")}
}
\description{
This function returns the neuron belonging to a given fc name as a neuronlist.
}
\examples{
fcneuronlist(c("fru-M-200145", "fru-M-200500"))
}

