% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot Single Cluster}
\usage{
plot_cluster(cluster_no, color = palette()[sample(length(palette()), 1)])
}
\arguments{
\item{cluster_no}{Integer of cluster number.}

\item{color}{Can set color; defaults to a random color in the existing palette.}
}
\description{
This function plots a single cluster. It takes input of cluster number as an integer from 1 to 809.
}
\examples{
plot_cluster(1, color = 'blue')
}

