% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_array.R
\name{plot_cluster_array}
\alias{plot_cluster_array}
\title{Plot Array of Clusters}
\usage{
plot_cluster_array(cluster_list, color = NULL)
}
\arguments{
\item{color}{Can set all clusters to a single color; defaults to plottting each cluster randomly among rainbow.}

\item{cluster_no}{array of cluster numbers}
}
\description{
This function plots an array of clusters. It takes input of array of cluster number as integers from 1 to 809.
}
\examples{
plot_cluster_array([1:809], color = 'blue')
}

