% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_cluster.R
\name{pop_cluster}
\alias{pop_cluster}
\title{Pop Cluster}
\usage{
pop_cluster(cluster_no)
}
\arguments{
\item{cluster_no}{cluster number}
}
\description{
This function pops a cluster from the viewer.
}
\examples{
pop_cluster(1)
}

